+++ src/core/admittance_matrix.jl
@@ -136,16 +136,41 @@ end
 Base.show(io::IO, x::AdmittanceMatrixInverse{<:Number}) = print(io, "AdmittanceMatrixInverse($(length(x.idx_to_bus)) buses, $(length(x.matrix)) entries)")
 
 
-"note, data should be a PowerModels network data model; only supports networks with exactly one refrence bus"
+"""
+    calc_susceptance_matrix_inv(data)
+
+Compute the inverse of the network's susceptance matrix.
+
+Note: `data`` should be a PowerModels network data model; only supports networks with exactly one refrence bus.
+
+While the susceptance matrix is sparse, its inverse it typically quite dense.
+This implementation first computes a sparse factorization, then recovers the (dense)
+    matrix inverse via backward substitution. This is more efficient
+    than directly computing a dense inverse with `LinearAlgebra.inv`.
+"""
 function calc_susceptance_matrix_inv(data::Dict{String,<:Any})
     #TODO check single connected component
-
     sm = calc_susceptance_matrix(data)
-
+    S  = sm.matrix
+    num_buses = length(sm.idx_to_bus)  # this avoids inactive buses
+    
     ref_bus = reference_bus(data)
-    sm_inv = calc_admittance_matrix_inv(sm, sm.bus_to_idx[ref_bus["index"]])
-
-    return sm_inv
+    ref_idx = sm.bus_to_idx[ref_bus["index"]]
+    if !(ref_idx > 0 && ref_idx <= num_buses)
+        Memento.error(_LOGGER, "invalid ref_idx in calc_susceptance_matrix_inv")
+    end
+    S[ref_idx, :] .= 0.0
+    S[:, ref_idx] .= 0.0
+    S[ref_idx, ref_idx] = 1.0
+    
+    F = LinearAlgebra.ldlt(Symmetric(S); check=false)
+    if !LinearAlgebra.issuccess(F)
+        Memento.error(_LOGGER, "Failed factorization in calc_susceptance_matrix_inv")
+    end
+    M = F \ Matrix(1.0I, num_buses, num_buses)
+    M[ref_idx, :] .= 0.0  # zero-out the row of the slack bus
+    
+    return AdmittanceMatrixInverse(sm.idx_to_bus, sm.bus_to_idx, ref_idx, M)
 end
 
 "calculates the inverse of the susceptance matrix"