+++ test/runtests.jl
@@ -45,6 +45,8 @@ include("common.jl")
 
     include("data.jl")
 
+    include("data-mc.jl")
+
     include("data-basic.jl")
 
     include("model.jl")